joining ctf_selection and ctf_discrepancy
