+++ config/templates/protocols.template
@@ -41,7 +41,7 @@ Protocols SPA = [
 		    {"tag": "protocol", "value": "ProtGctf", "text": "default"},
 			{"tag": "section", "text": "more", "openItem": "False", "children": [
 				{"tag": "protocol", "value": "ProtCTFAssign",           "text": "default"},
-				{"tag": "protocol", "value": "XmippProtCTFDiscrepancy", "text": "default"},
+				{"tag": "protocol", "value": "XmippProtCTFConsensus", "text": "default"},
 				{"tag": "protocol", "value": "ProtRelionExportCtf", "text": "default"},
 				{"tag": "protocol", "value": "XmippProtCTFSelection", "text": "default"}
 			]}
