+++ pyworkflow/em/metadata/constants.py
@@ -95,6 +95,7 @@ MDL_CONTINUOUS_GRAY_A = xmipp.MDL_CONTINUOUS_GRAY_A
 MDL_CONTINUOUS_GRAY_B = xmipp.MDL_CONTINUOUS_GRAY_B
 MDL_CONTINUOUS_X = xmipp.MDL_CONTINUOUS_X
 MDL_CONTINUOUS_Y = xmipp.MDL_CONTINUOUS_Y
+MDL_CONSENSUS_RESOLUTION = xmipp.MDL_CONSENSUS_RESOLUTION
 MDL_COST = xmipp.MDL_COST
 MDL_COUNT = xmipp.MDL_COUNT
 MDL_COUNT2 = xmipp.MDL_COUNT2
