+++ pyworkflow/em/packages/xmipp3/protocol_ctf_consensus.py
@@ -1,7 +1,9 @@
 # **************************************************************************
 # *
-# * Authors:     Roberto Marabini (roberto@cnb.csic.es)
-# *              Amaya Jimenez    (ajimenez@cnb.csic.es)
+# * Authors:     J.M. De la Rosa Trevin (jmdelarosa@cnb.csic.es)
+# *              Roberto Marabini (roberto@cnb.csic.es)
+# *              Tomas Majtner (tmajtner@cnb.csic.es)  -- streaming version
+# *              Amaya Jimenez (ajimenez@cnb.csic.es)
 # *
 # * Unidad de  Bioinformatica of Centro Nacional de Biotecnologia , CSIC
 # *
@@ -25,36 +27,38 @@
 # *
 # **************************************************************************
 
-
-from os.path import getmtime, exists
+import os
+import convert
+import xmipp
 from datetime import datetime
-
-import pyworkflow.em as em
+from pyworkflow.em.data import SetOfCTF
+from pyworkflow.object import Set, Float
 import pyworkflow.protocol.params as params
+import pyworkflow.em as em
 import pyworkflow.utils as pwutils
-
-from pyworkflow.object import Set
+from pyworkflow.em.metadata import Row, MetaData
 from pyworkflow.protocol.constants import (STATUS_NEW)
 
 
-class XmippProtCTFSelection(em.ProtCTFMicrographs):
+class XmippProtCTFConsensus(em.ProtCTFMicrographs):
     """
-    Protocol to make a selection of meaningful CTFs in basis of the defocus
-    values, the astigmatism, and the resolution.
+    Protocol to estimate the agreement between different estimation of the CTF
+    for the same set of micrographs. The algorithm assumes that two CTF are
+    consistent if the phase (wave aberration function) of the two CTFs are
+    closer than 90 degrees. The reported resolution is the resolution at
+    which the two CTF phases differ in 90 degrees.
     """
-    _label = 'ctf selection'
+    _label = 'ctf consensus'
 
     def __init__(self, **args):
         em.ProtCTFMicrographs.__init__(self, **args)
+        self._freqResol = {}
+        self.stepsExecutionMode = params.STEPS_SERIAL
 
-    # --------------------------- DEFINE param functions ------------------
     def _defineParams(self, form):
         form.addSection(label='Input')
-        # Read a ctf estimation
-        form.addParam('inputCTF', params.PointerParam,
-                      pointerClass='SetOfCTF',
-                      label="Input CTF",
-                      important=True,
+        form.addParam('inputCTF', params.PointerParam, pointerClass='SetOfCTF',
+                      label="Input CTF", important=True,
                       help='Select the estimated CTF to evaluate')
         form.addParam('useDefocus', params.BooleanParam, default=True,
                       label='Use Defocus for selection',
@@ -123,87 +127,173 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
                       default=0.3, label='Min')
         line.addParam('maxCritNonAstigmaticValidity', params.FloatParam,
                       default=25, label='Max')
+        form.addParam('calculateConsensus', params.BooleanParam, default=False,
+                      label='Calculate Consensus Resolution',
+                      help='Option for calculating consensus resolution.')
+        form.addParam('inputCTF2', params.PointerParam,
+                      pointerClass='SetOfCTF', condition="calculateConsensus",
+                      label="target CTF",
+                      help='CTF to be compared with reference CTF')
+        form.addParam('minConsResol', params.FloatParam,
+                      condition="calculateConsensus", default=5.0,
+                      label='Minimum consensus resolution in angstroms.')
+        form.addParallelSection(threads=0, mpi=0)
 
-    # --------------------------- INSERT steps functions ------------------
+# --------------------------- INSERT steps functions -------------------------
     def _insertAllSteps(self):
-
-        fDeps = []
+        self.finished = False
         self.insertedDict = {}
+        self.processedDict = []
+        self.outputDict = []
+        self.allCtf1 = []
+        self.allCtf2 = []
+        if self.calculateConsensus:
+            ctfSteps = self._checkNewInput()
+        else:
+            ctfSteps = self._insertNewSelectionSteps(self.insertedDict,
+                                                     self.inputCTF.get())
+        self._insertFunctionStep('createOutputStep',
+                                 prerequisites=ctfSteps, wait=True)
 
-        fDeps = self._insertNewSelectionSteps(self.insertedDict,
-                                              self.inputCTF.get())
-        # For the streaming mode, the steps function have a 'wait' flag
-        # that can be turned on/off. For example, here we insert the
-        # createOutputStep but it wait=True, which means that can not be
-        # executed until it is set to False
-        # (when the input micrographs stream is closed)
-        self._insertFunctionStep('createOutputStep', prerequisites=fDeps,
-                                 wait=True)
+    def createOutputStep(self):
+        pass
+
+    def _getFirstJoinStepName(self):
+        # This function will be used for streaming, to check which is
+        # the first function that need to wait for all ctfs
+        # to have completed, this can be overriden in subclasses
+        # (e.g., in Xmipp 'sortPSDStep')
+        return 'createOutputStep'
+
+    def _getFirstJoinStep(self):
+        for s in self._steps:
+            if s.funcName == self._getFirstJoinStepName():
+                return s
+        return None
+
+    def _insertNewCtfsSteps(self, SetOfCtf1, SetOfCtf2, insDict):
+        deps = []
+        discrepId = self._insertFunctionStep("_computeCTFDiscrepancy",
+                                             SetOfCtf1, SetOfCtf2,
+                                             prerequisites=[])
+        deps.append(discrepId)
+        if (len(SetOfCtf1) > len(SetOfCtf2)):
+            SetOfCtf = SetOfCtf2
+        else:
+            SetOfCtf = SetOfCtf1
+        for ctf in SetOfCtf:
+            ctfId = ctf.getObjId()
+            if ctfId not in insDict:
+                stepId = self._insertFunctionStep('_selectCTF', ctfId,
+                                                  prerequisites=[discrepId])
+                deps.append(stepId)
+                insDict[ctfId] = stepId
+
+        return deps
 
     def _insertNewSelectionSteps(self, insertedDict, inputCtfs):
-        """ Insert steps to process new ctfs (from streaming)
-        Params:
-            insertedDict: contains already processed ctfs
-            inputCtfs: input ctfs set to be check
-        """
         deps = []
         # For each ctf insert the step to process it
         for ctf in inputCtfs:
             ctfId = ctf.getObjId()
             if ctfId not in insertedDict:
-                stepId = self._insertCtfSelectionStep(ctfId)
+                stepId = self._insertFunctionStep('_selectCTF', ctfId,
+                                                  prerequisites=[])
                 deps.append(stepId)
                 insertedDict[ctfId] = stepId
         return deps
 
-    def _insertCtfSelectionStep(self, ctfId):
-        """ Insert the processMovieStep for a given movie. """
-        # Note1: At this point is safe to pass the movie, since this
-        # is not executed in parallel, here we get the params
-        # to pass to the actual step that is gone to be executed later on
-        # Note2: We are serializing the Movie as a dict that can be passed
-        # as parameter for a functionStep
-        stepId = self._insertFunctionStep('_selectCTF', ctfId,
-                                          prerequisites=[])
-        return stepId
-
-    # --------------------------- STEPS functions ------------------------
     def _stepsCheck(self):
-
-        # check if there are new ctfs and process them
         self._checkNewInput()
         self._checkNewOutput()
 
     def _checkNewInput(self):
-        """ Check if there are new ctf to be processed and add the necessary
-        steps."""
-        ctfFile = self.inputCTF.get().getFileName()
+        if self.calculateConsensus:
+            # Check if there are new ctf to process from the input set
+            ctfsFile1 = self.inputCTF.get().getFileName()
+            ctfsFile2 = self.inputCTF2.get().getFileName()
+            self.lastCheck = getattr(self, 'lastCheck', datetime.now())
+            mTime = max(datetime.fromtimestamp(os.path.getmtime(ctfsFile1)),
+                        datetime.fromtimestamp(os.path.getmtime(ctfsFile2)))
+            # If the input movies.sqlite have not changed since our last check,
+            # it does not make sense to check for new input data
+            if self.lastCheck > mTime and hasattr(self, 'SetOfCtf1'):
+                return None
+            ctfsSet1 = SetOfCTF(filename=ctfsFile1)
+            ctfsSet2 = SetOfCTF(filename=ctfsFile2)
+            ctfsSet1.loadAllProperties()
+            ctfsSet2.loadAllProperties()
+            if len(self.allCtf1) > 0:
+                newCtf1 = [ctf.clone() for ctf in
+                           ctfsSet1.iterItems(orderBy='creation',
+                                              where='creation>"' + str(
+                                                  self.checkCtf1) + '"')]
+            else:
+                newCtf1 = [ctf.clone() for ctf in ctfsSet1]
+            self.allCtf1 = self.allCtf1 + newCtf1
+            if len(newCtf1) > 0:
+                for ctf in ctfsSet1.iterItems(orderBy='creation',
+                                              direction='DESC'):
+                    self.checkCtf1 = ctf.getObjCreation()
+                    break
+            if len(self.allCtf2) > 0:
+                newCtf2 = [ctf.clone() for ctf in
+                           ctfsSet2.iterItems(orderBy='creation',
+                                              where='creation>"' + str(
+                                                  self.checkCtf2) + '"')]
+            else:
+                newCtf2 = [ctf.clone() for ctf in ctfsSet2]
+            self.allCtf2 = self.allCtf2 + newCtf2
+            if len(newCtf2) > 0:
+                for ctf in ctfsSet2.iterItems(orderBy='creation',
+                                              direction='DESC'):
+                    self.checkCtf2 = ctf.getObjCreation()
+                    break
+            self.lastCheck = datetime.now()
+            self.isStreamClosed = ctfsSet1.isStreamClosed() and \
+                                  ctfsSet2.isStreamClosed()
+            ctfsSet1.close()
+            ctfsSet2.close()
+            outputStep = self._getFirstJoinStep()
+            if len(set(self.allCtf1)) > len(set(self.processedDict)) and \
+               len(set(self.allCtf2)) > len(set(self.processedDict)):
+                fDeps = self._insertNewCtfsSteps(ctfsSet1, ctfsSet2,
+                                                 self.insertedDict)
+                if outputStep is not None:
+                    outputStep.addPrerequisites(*fDeps)
+                self.updateSteps()
+        else:
+            ctfFile = self.inputCTF.get().getFileName()
+            now = datetime.now()
+            self.lastCheck = getattr(self, 'lastCheck', now)
+            mTime = datetime.fromtimestamp(os.path.getmtime(ctfFile))
+            self.debug('Last check: %s, modification: %s'
+                       % (pwutils.prettyTime(self.lastCheck),
+                          pwutils.prettyTime(mTime)))
 
-        now = datetime.now()
-        self.lastCheck = getattr(self, 'lastCheck', now)
-        mTime = datetime.fromtimestamp(getmtime(ctfFile))
-        self.debug('Last check: %s, modification: %s'
-                   % (pwutils.prettyTime(self.lastCheck),
-                      pwutils.prettyTime(mTime)))
+            # Open input ctfs.sqlite and close it as soon as possible
+            ctfSet = self._loadInputCtfSet()
+            self.isStreamClosed = ctfSet.isStreamClosed()
+            self.allCtf1 = [m.clone() for m in ctfSet]
+            ctfSet.close()
 
-        # Open input ctfs.sqlite and close it as soon as possible
-        self._loadInputList()
-        # If the input ctfs.sqlite have not changed since our last check,
-        # it does not make sense to check for new input data
-        if self.lastCheck > mTime and hasattr(self, 'listOfCtf'):
-            return None
+            # If the input ctfs.sqlite have not changed since our last check,
+            # it does not make sense to check for new input data
+            if self.lastCheck > mTime and hasattr(self, 'allCtf1'):
+                return None
 
-        self.lastCheck = now
-        newCtf = any(ctf.getObjId() not in
-                     self.insertedDict for ctf in self.listOfCtf)
-        outputStep = self._getFirstJoinStep()
+            self.lastCheck = now
+            newCtf = any(ctf.getObjId() not in
+                         self.insertedDict for ctf in self.allCtf1)
+            outputStep = self._getFirstJoinStep()
 
-        if newCtf:
-            fDeps = self._insertNewSelectionSteps(self.insertedDict,
-                                                  self.listOfCtf)
-            if outputStep is not None:
-                outputStep.addPrerequisites(*fDeps)
-            self.updateSteps()
+            if newCtf:
+                fDeps = self._insertNewSelectionSteps(self.insertedDict,
+                                                      self.allCtf1)
+                if outputStep is not None:
+                    outputStep.addPrerequisites(*fDeps)
+                self.updateSteps()
+
 
     def _checkNewOutput(self):
         """ Check for already selected CTF and update the output set. """
@@ -227,8 +317,8 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
 
         # We have finished when there is not more input ctf (stream closed)
         # and the number of processed ctf is equal to the number of inputs
-        self.finished = (self.isStreamClosed == Set.STREAM_CLOSED and
-                         allDone == len(self.listOfCtf))
+        self.finished = (self.isStreamClosed and allDone == len(self.allCtf1))
+
         streamMode = Set.STREAM_CLOSED if self.finished else Set.STREAM_OPEN
 
         # reading the outputs
@@ -254,6 +344,11 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
                 ctf.setEnabled(self._getEnable(ctfId))
                 mic.setEnabled(self._getEnable(ctfId))
 
+                if self.calculateConsensus:
+                    ctf.setConsensusResolution(self._freqResol[ctfId])
+                    ctf._discrepancy_astigmatism = Float(ctf.getDefocusU() -
+                                                        ctf.getDefocusV())
+
                 ctfSet.append(ctf)
                 micSet.append(mic)
                 self._writeDoneListAccepted(ctfId)
@@ -264,6 +359,10 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
             inputCtfSet = self._loadInputCtfSet()
             for ctfId in newDoneDiscarded:
                 ctf = inputCtfSet[ctfId].clone()
+                if self.calculateConsensus:
+                    ctf.setConsensusResolution(self._freqResol[ctfId])
+                    ctf._discrepancy_astigmatism = Float(ctf.getDefocusU() -
+                                                        ctf.getDefocusV())
                 mic = ctf.getMicrograph().clone()
                 micSetDiscarded.append(mic)
                 ctfSetDiscarded.append(ctf)
@@ -277,17 +376,17 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
             # so we exit from the function here
             return
 
-        if (exists(self._getPath('ctfs.sqlite'))):
+        if (os.path.exists(self._getPath('ctfs.sqlite'))):
             self._updateOutputSet('outputCTF', ctfSet, streamMode)
             self._updateOutputSet('outputMicrographs', micSet, streamMode)
         # AJ new subsets with discarded ctfs
-        if (exists(self._getPath('ctfsDiscarded.sqlite'))):
+        if (os.path.exists(self._getPath('ctfsDiscarded.sqlite'))):
             self._updateOutputSet('outputCTFDiscarded',
                                   ctfSetDiscarded, streamMode)
             self._updateOutputSet('outputMicrographsDiscarded',
                                   micSetDiscarded, streamMode)
 
-        if (exists(self._getPath('ctfs.sqlite'))):
+        if (os.path.exists(self._getPath('ctfs.sqlite'))):
             if firstTimeAccepted:
                 # define relation just once
                 self._defineSourceRelation(
@@ -300,7 +399,7 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
                 micSet.close()
 
         # AJ new subsets with discarded ctfs
-        if (exists(self._getPath('ctfsDiscarded.sqlite'))):
+        if (os.path.exists(self._getPath('ctfsDiscarded.sqlite'))):
             if firstTimeDiscarded:
                 self._defineSourceRelation(
                     self.inputCTF.get().getMicrographs(), micSetDiscarded)
@@ -316,17 +415,71 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
             if outputStep and outputStep.isWaiting():
                 outputStep.setStatus(STATUS_NEW)
 
-        if (exists(self._getPath('ctfs.sqlite'))):
+        if (os.path.exists(self._getPath('ctfs.sqlite'))):
             ctfSet.close()
             micSet.close()
         # AJ new subsets with discarded ctfs
-        if (exists(self._getPath('ctfsDiscarded.sqlite'))):
+        if (os.path.exists(self._getPath('ctfsDiscarded.sqlite'))):
             micSetDiscarded.close()
             ctfSetDiscarded.close()
 
-    def createOutputStep(self):
-        # Do nothing now, the output should be ready.
-        pass
+
+    def _loadOutputSet(self, SetClass, baseName):
+        """
+        Load the output set if it exists or create a new one.
+        """
+        setFile = self._getPath(baseName)
+
+        if os.path.exists(setFile):
+            outputSet = SetClass(filename=setFile)
+            if (outputSet.__len__() is 0):
+                pwutils.path.cleanPath(setFile)
+
+        if os.path.exists(setFile):
+            outputSet = SetClass(filename=setFile)
+            outputSet.loadAllProperties()
+            outputSet.enableAppend()
+        else:
+            outputSet = SetClass(filename=setFile)
+            outputSet.setStreamState(outputSet.STREAM_OPEN)
+
+        micSet = self.inputCTF.get().getMicrographs()
+
+        if isinstance(outputSet, em.SetOfMicrographs):
+            outputSet.copyInfo(micSet)
+        elif isinstance(outputSet, em.SetOfCTF):
+            outputSet.setMicrographs(micSet)
+        return outputSet
+
+    def _ctfToMd(self, ctf, ctfMd):
+        """ Write the proper metadata for Xmipp from a given CTF """
+        ctfMd.clear()
+        ctfRow = Row()
+        convert.ctfModelToRow(ctf, ctfRow)
+        convert.micrographToRow(ctf.getMicrograph(), ctfRow,
+                                alignType=convert.ALIGN_NONE)
+        ctfRow.addToMd(ctfMd)
+
+    def _computeCTFDiscrepancy(self, method1, method2):
+        # TODO must be same micrographs
+        # move to a single step, each step takes 5 sec while the function
+        # takes 0.03 sec
+        # convert to md
+        md1 = MetaData()
+        md2 = MetaData()
+
+        for ctf1 in method1:  # reference CTF
+            ctfId = ctf1.getObjId()
+            if ctfId in self.processedDict:
+                continue
+            for ctf2 in method2:
+                ctfId2 = ctf2.getObjId()
+                if ctfId2 != ctfId:
+                    continue
+                self.processedDict.append(ctfId)
+                self._ctfToMd(ctf1, md1)
+                self._ctfToMd(ctf2, md2)
+                self._freqResol[ctfId] = xmipp.errorMaxFreqCTFs2D(md1, md2)
 
     def _selectCTF(self, ctfId):
         # Depending on the flags selected by the user, we set the values of
@@ -348,6 +501,10 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
         firstCondition = ((defocusU < minDef) or (defocusU > maxDef) or
                           (defocusV < minDef) or (defocusV > maxDef) or
                           (astigm > maxAstig) or (resol > minResol))
+
+        if self.calculateConsensus:
+            firstCondition = firstCondition or (
+            self.minConsResol < self._freqResol[ctfId])
 
         secondCondition = False
         if (self.useCritXmipp and usingXmipp):
@@ -384,24 +541,44 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
                 with open(fn, 'a') as f:
                     f.write('%d F\n' % ctf.getObjId())
 
-    # --------------------------- INFO functions --------------------------
+    def _readDoneList(self):
+        """ Read from a file the id's of the items that have been done. """
+        doneFile = self._getAllDone()
+        doneList = []
+        # Check what items have been previously done
+        if os.path.exists(doneFile):
+            with open(doneFile) as f:
+                doneList += [int(line.strip()) for line in f]
+        return doneList
+
+    def _getAllDone(self):
+        return self._getExtraPath('DONE_all.TXT')
+
+    def _writeDoneList(self, partList):
+        """ Write to a text file the items that have been done. """
+        with open(self._getAllDone(), 'a') as f:
+            for part in partList:
+                f.write('%d\n' % part.getObjId())
+
+    def _citations(self):
+        return ['Marabini2014a']
+
     def _summary(self):
         message = []
+        for i, ctfs in enumerate([self.inputCTF, self.inputCTF2]):
+            protocol = self.getMapper().getParent(ctfs.get())
+            message.append("Method %d: %s" % (i+1, protocol.getClassLabel()))
         return message
-
-    def _methods(self):
-        # nothing here
-        pass
 
     def _validate(self):
         """ The function of this hook is to add some validation before the
         protocol is launched to be executed. It should return a list of
         errors. If the list is empty the protocol can be executed.
         """
-        message = []
-        return message
-
-    # --------------------------- UTILS functions ---------------------
+        # same micrographs in both CTF??
+        errors = []
+        # Add some errors if input is not valid
+        return errors
 
     def usingXmipp(self, ctf):
         return ctf.hasAttribute('_xmipp_ctfCritfirstZeroRatio')
@@ -412,15 +589,6 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
             return resolution
         else:
             return 0
-        # if ctf.hasAttribute('_ctffind4_ctfResolution'):
-        #     return ctf._ctffind4_ctfResolution.get(), False
-        # elif ctf.hasAttribute('_gctf_ctfResolution'):
-        #     return ctf._gctf_ctfResolution.get(), False
-        # elif ctf.hasAttribute('_xmipp_ctfCritMaxFreq'):
-        #     return ctf._xmipp_ctfCritMaxFreq.get(), True
-        # else:
-        #     # if 0, the protocol does not select by ctf resolution.
-        #     return 0
 
     def _readDoneListDiscarded(self):
         """ Read from a text file the id's of the items
@@ -428,10 +596,9 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
         DiscardedFile = self._getDiscardedDone()
         DiscardedList = []
         # Check what items have been previously done
-        if exists(DiscardedFile):
+        if os.path.exists(DiscardedFile):
             with open(DiscardedFile) as f:
                 DiscardedList += [int(line.strip()) for line in f]
-
         return DiscardedList
 
     def _readDoneListAccepted(self):
@@ -440,10 +607,9 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
         AcceptedFile = self._getAcceptedDone()
         AcceptedList = []
         # Check what items have been previously done
-        if exists(AcceptedFile):
+        if os.path.exists(AcceptedFile):
             with open(AcceptedFile) as f:
                 AcceptedList += [int(line.strip()) for line in f]
-
         return AcceptedList
 
     def _writeDoneListDiscarded(self, ctfId):
@@ -470,42 +636,6 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
     def _getCtfSelecFileDiscarded(self):
         return self._getExtraPath('selection-ctf-discarded.txt')
 
-    def _loadInputList(self):
-        """ Load the input set of ctfs and create a list. """
-        ctfSet = self._loadInputCtfSet()
-        self.isStreamClosed = ctfSet.getStreamState()
-        self.listOfCtf = [m.clone() for m in ctfSet]
-        ctfSet.close()
-        self.debug("Closed db.")
-
-    def _loadOutputSet(self, SetClass, baseName):
-        """
-        Load the output set if it exists or create a new one.
-        """
-        setFile = self._getPath(baseName)
-
-        if exists(setFile):
-            outputSet = SetClass(filename=setFile)
-            if(outputSet.__len__() is 0):
-                pwutils.path.cleanPath(setFile)
-
-        if exists(setFile):
-            outputSet = SetClass(filename=setFile)
-            outputSet.loadAllProperties()
-            outputSet.enableAppend()
-        else:
-            outputSet = SetClass(filename=setFile)
-            outputSet.setStreamState(outputSet.STREAM_OPEN)
-
-        micSet = self.inputCTF.get().getMicrographs()
-
-        if isinstance(outputSet, em.SetOfMicrographs):
-            outputSet.copyInfo(micSet)
-        elif isinstance(outputSet, em.SetOfCTF):
-            outputSet.setMicrographs(micSet)
-
-        return outputSet
-
     def _readtCtfId(self, accepted):
         if accepted:
             fn = self._getCtfSelecFileAccepted()
@@ -513,16 +643,15 @@ class XmippProtCTFSelection(em.ProtCTFMicrographs):
             fn = self._getCtfSelecFileDiscarded()
         ctfList = []
         # Check what items have been previously done
-        if exists(fn):
+        if os.path.exists(fn):
             with open(fn) as f:
                 ctfList += [int(line.strip().split()[0]) for line in f]
         return ctfList
 
     def _getEnable(self, ctfId):
         fn = self._getCtfSelecFileAccepted()
-        ctfList = []
         # Check what items have been previously done
-        if exists(fn):
+        if os.path.exists(fn):
             with open(fn) as f:
                 for line in f:
                     if ctfId == int(line.strip().split()[0]):
