+++ pyworkflow/em/packages/xmipp3/viewer.py
@@ -43,7 +43,7 @@ from protocol_cl2d_align import XmippProtCL2DAlign
 from protocol_cl2d import XmippProtCL2D
 from protocol_compare_reprojections import XmippProtCompareReprojections
 from protocol_compare_angles import XmippProtCompareAngles
-from protocol_ctf_discrepancy import XmippProtCTFDiscrepancy
+from protocol_ctf_consensus import XmippProtCTFConsensus
 from protocol_extract_particles import XmippProtExtractParticles
 from protocol_extract_particles_pairs import XmippProtExtractParticlesPairs
 from protocol_helical_parameters import XmippProtHelicalParameters
