+++ pyworkflow/tests/em/workflows/test_workflow_xmipp_ctf_consensus.py
@@ -31,7 +31,7 @@ import os
 from itertools import izip
 
 from pyworkflow.em.data import SetOfCTF, CTFModel, Micrograph, SetOfMicrographs
-from pyworkflow.em.packages.xmipp3 import XmippProtCTFDiscrepancy
+from pyworkflow.em.packages.xmipp3 import XmippProtCTFConsensus
 from pyworkflow.em.protocol import ProtImportMicrographs, ProtImportCTF
 from pyworkflow.object import PointerList, Pointer
 from test_workflow import TestWorkflow
@@ -40,7 +40,7 @@ from pyworkflow.em import ImageHandler
 import xmipp
 
 
-class TestXmippCTFDiscrepancyBase(TestWorkflow):
+class TestXmippCTFConsensusBase(TestWorkflow):
     @classmethod
     def setUpClass(cls):
         tests.setupTestProject(cls)
@@ -53,7 +53,7 @@ class TestXmippCTFDiscrepancyBase(TestWorkflow):
 
         return ctf
 
-    def testCtfdiscrepancyWorkflow(self):
+    def testCtfConsensusWorkflow(self):
         # create one micrograph set
         fnMicSet = self.proj.getTmpPath("mics.sqlite")
         fnMic = self.proj.getTmpPath("mic.mrc")
@@ -134,13 +134,13 @@ class TestXmippCTFDiscrepancyBase(TestWorkflow):
         self.launchProtocol(protCTF1)
         self.launchProtocol(protCTF2)
 
-        # launch CTF discrepancy protocol
-        protCtfDiscrepancy = self.newProtocol(XmippProtCTFDiscrepancy)
-        protCtfDiscrepancy.inputCTF1.set(protCTF1.outputCTF)
-        protCtfDiscrepancy.inputCTF2.set(protCTF2.outputCTF)
-        protCtfDiscrepancy.setObjLabel('ctf discrepancy')
-        self.launchProtocol(protCtfDiscrepancy)
-        ctf0 = protCtfDiscrepancy.outputCTF.getFirstItem()
+        # launch CTF consensus protocol
+        protCtfConsensus = self.newProtocol(XmippProtCTFConsensus)
+        protCtfConsensus.inputCTF1.set(protCTF1.outputCTF)
+        protCtfConsensus.inputCTF2.set(protCTF2.outputCTF)
+        protCtfConsensus.setObjLabel('ctf consensus')
+        self.launchProtocol(protCtfConsensus)
+        ctf0 = protCtfConsensus.outputCTF.getFirstItem()
         resolution = int(ctf0.getResolution())
         defocusU = int(ctf0.getDefocusU())
         self.assertEqual(resolution, 2)
