+++ software/em/xmipp/libraries/bindings/python/python_constants.cpp
@@ -120,6 +120,7 @@ void addLabels(PyObject * dict)
     ADD_CONST(MDL_COST_PERCENTILE);
     ADD_CONST(MDL_COUNT);
     ADD_CONST(MDL_COUNT2);
+    ADD_CONST(MDL_CONSENSUS_RESOLUTION);
     ADD_CONST(MDL_CONTINUOUS_GRAY_A);
     ADD_CONST(MDL_CONTINUOUS_GRAY_B);
     ADD_CONST(MDL_CONTINUOUS_X);
